wzhang: bug fixed

+++ src/main/java/NullValueHandler.java
@@ -49,7 +49,7 @@ public class NullValueHandler {
 
     public static void main(String[] args) throws Exception{
         Configuration conf = new Configuration();
-        conf.setFloat("movies", Integer.parseInt(args[2]));
+        conf.setInt("movies", Integer.parseInt(args[2]));
 
         Job job = Job.getInstance(conf);
         job.setMapperClass(NullMapper.class);
